# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Leerr API
  version: "0.2"
  description: |
    Same-origin JSON API for the Leerr web app.

    Authentication: web sessions use an HttpOnly, SameSite=Strict cookie
    (`__Host-leerr` over HTTPS) and must send `X-CSRF-Token` on every
    non-GET request. Native sessions send `Authorization: Bearer <token>`
    and need no CSRF token. Each session kind only works through its own channel.

    Every error has the body `{"error": {"code": "...", "message": "..."}}`;
    messages are written for end users.
servers:
  - url: /
security:
  - cookie: []
  - bearer: []
...
paths:
...
  /api/v1/library:
    get:
      tags: [Library]
      summary: A page of albums visible to the user's Jellyfin account
      parameters:
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 60 } }
        - { name: q, in: query, schema: { type: string, maxLength: 200 } }
      responses:
        "200":
          description: Albums
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Album" } }
                  total: { type: integer }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/albums/{id}:
    get:
      tags: [Library]
      summary: An album and its tracks
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Album
          content:
            application/json:
              schema:
                type: object
                required: [album, tracks]
                properties:
                  album: { $ref: "#/components/schemas/Album" }
                  tracks: { type: array, items: { $ref: "#/components/schemas/Track" } }
  /api/v1/artwork/{id}:
    get:
      tags: [Library]
      summary: Album artwork through the user's Jellyfin access
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Image
          content:
            image/*: { schema: { type: string, format: binary } }
  /api/v1/stream-tickets:
    post:
      tags: [Library]
      summary: Authorise playback of one track
      description: |
        Returns a URL whose unguessable path is the credential, for media
        elements that cannot send headers. It admits new requests for 6 hours;
        a stream that has started is never cut off by expiry.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [trackID], properties: { trackID: { type: string } } }
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                type: object
                required: [url, expiresAt]
                properties:
                  url: { type: string }
                  expiresAt: { type: string, format: date-time }
  /api/v1/streams/{ticket}:
    get:
      tags: [Library]
      summary: The original audio file, with Range support
      security: []
      parameters:
        - { name: ticket, in: path, required: true, schema: { type: string } }
        - { name: Range, in: header, schema: { type: string } }
        - { name: If-Range, in: header, schema: { type: string } }
      responses:
        "200": { description: Whole file, content: { audio/*: { schema: { type: string, format: binary } } } }
        "206": { description: Requested range, content: { audio/*: { schema: { type: string, format: binary } } } }
        "401": { $ref: "#/components/responses/Error" }
        "416": { description: Range not satisfiable }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/discover/search:
    get:
      tags: [Discovery]
      summary: Search MusicBrainz for albums and artists
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 1, maxLength: 200 } }
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/CandidatePage"
                  - type: object
                    required: [artists]
                    properties:
                      artists:
                        type: array
                        items:
                          type: object
                          required: [id, name, disambiguation, country, type]
                          properties:
                            id: { type: string, format: uuid }
                            name: { type: string }
                            disambiguation: { type: string }
                            country: { type: string }
                            type: { type: string }
  /api/v1/discover/artists/{id}/albums:
    get:
      tags: [Discovery]
      summary: An artist's albums from MusicBrainz, 25 per page
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, maximum: 10000 } }
      responses:
        "200":
          description: Albums
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CandidatePage" }
  /api/v1/discover/release-groups/{id}/editions:
    get:
      tags: [Discovery]
      summary: Editions (MusicBrainz releases) of an album, oldest first
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Editions
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, title, date, country, formats, tracks]
                      properties:
                        id: { type: string, format: uuid }
                        title: { type: string }
                        date: { type: [string, "null"] }
                        country: { type: [string, "null"] }
                        formats: { type: string }
                        tracks: { type: [integer, "null"] }
  /api/v1/discover/recommendations:
    get:
      tags: [Discovery]
      summary: Suggested albums not already owned or requested
      description: Withheld (409/5xx) when the user's Jellyfin library cannot be read to filter owned albums.
      parameters:
        - { name: refresh, in: query, schema: { type: string, enum: ["1"] } }
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: object
                required: [items, source, emptyReason]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Candidate" } }
                  source: { type: string, enum: [lastfm, musicbrainz] }
                  emptyReason: { type: [string, "null"], enum: [no_candidates, all_excluded, null] }
...
components:
  securitySchemes:
    cookie: { type: apiKey, in: cookie, name: __Host-leerr }
    bearer: { type: http, scheme: bearer }
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
  schemas:
...
    Album:
      type: object
      required: [id, title, artist, year, releaseGroupMBID, releaseMBID]
      properties:
        id: { type: string }
        title: { type: string }
        artist: { type: string }
        year: { type: [integer, "null"] }
        releaseGroupMBID: { type: [string, "null"] }
        releaseMBID: { type: [string, "null"] }
    Track:
      type: object
      required: [id, title, artist, disc, number, duration, codec, sampleRate, bitDepth]
      properties:
        id: { type: string }
        title: { type: string }
        artist: { type: string }
        disc: { type: [integer, "null"] }
        number: { type: [integer, "null"] }
        duration: { type: [number, "null"], description: Seconds. }
        codec: { type: [string, "null"] }
        sampleRate: { type: [number, "null"] }
        bitDepth: { type: [integer, "null"] }
    Candidate:
      type: object
      required: [id, title, artist, artistMBID, year, coverUrl, status]
      properties:
        id: { type: string, format: uuid, description: MusicBrainz release group. }
        title: { type: string }
        artist: { type: string }
        artistMBID: { type: string, format: uuid }
        year: { type: [string, "null"] }
        coverUrl: { type: [string, "null"] }
        status: { type: [string, "null"], enum: [available, requested, null] }
    CandidatePage:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Candidate" } }
        total: { type: integer }